add formatting/linting, move to private dir
