+++ .pre-commit-config.yaml
@@ -1,6 +1,7 @@
 repos:
-  - repo: https://github.com/psf/black
-    rev: 24.3.0
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.5.7
     hooks:
-      - id: black
-        language_version: python3.12
+      - id: ruff # linter
+        args: ["--fix"]
+      - id: ruff-format
