+++ justfile
@@ -6,9 +6,25 @@ _default:
 dtest TEST:
   python -m debugpy --listen 0.0.0.0:5680 --wait-for-client -m pytest tests/{{TEST}}
 
+# debug an individual test with -k flag
+dktest ARG:
+  python -m debugpy --listen 0.0.0.0:5680 --wait-for-client -m pytest -k {{ARG}}
+
 # debug a script
 dscript FILE:
   python -m debugpy --listen 0.0.0.0:5680 --wait-for-client {{FILE}}
+
+# format the repo
+format:
+  poetry run ruff format
+
+# lint the repo
+lint:
+  poetry run ruff check
+
+# lint the repo (+ auto-fix)
+lintfix:
+  poetry run ruff check --fix
 
 # Run all test suites
 tests:
@@ -17,3 +33,7 @@ tests:
 # run an individual test under tests/
 test TEST:
   pytest tests/{{TEST}}
+
+# run an individual test with -k flag
+ktest ARG:
+  pytest -k {{ARG}}
