+++ tests/test_cons.py
@@ -160,6 +160,29 @@ def test_prepend_as_dunder(input, expected, request):
     assert input == e
 
 
+# def test_yo():
+#   class Nill:
+#       def __repr__(self):
+#           return "Nil"
+
+#   class Conss:
+#       def __init__(self, head, tail):
+#           self.head = head
+#           self.tail = tail
+
+#       def __repr__(self):
+#           return f"Conss({self.head}, {self.tail})"
+
+#       def __rlshift__(self, other):
+#           return Conss(other, self)
+
+#   result = 1 << 2 << 3 << Nill()  # Creates Cons(1, Cons(2, Cons(3, Nil())))
+#   print(result)  # Output: Cons(1, Cons(2, Cons(3, Nil())))
+#   # nil = Nil()
+#   # result = 2 << 1 << nil
+#   # print(result)
+
+
 @P.autodetect_parameters()
 @P.case(
     name="Nil",
@@ -512,3 +535,13 @@ def test_flatten(input, expected):
 def test_flat_map(original, expected):
     triple = lambda n: CList.from_iterable([n * 3, n * 3])
     assert original.flat_map(triple) == expected
+
+
+@pytest.mark.skip
+def test_prepend_tailrec():
+    n = 10000  # recursion limit
+    lst = Nil()
+    while n > 0:
+        lst = n << lst
+        n -= 1
+    assert len(lst)
